hilder_gv/requirements.txt
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
idna==2.6
lxml==4.1.1
py==1.5.2
pymongo==3.6.0
PyYAML==3.12
requests==2.18.4
retry==0.9.2
urllib3==1.22
